pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.4",
    "matplotlib>=3.10.8",
    "mplfinance>=0.12.10b0",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot[job-queue]>=22.7",
    "requests>=2.33.0",
]
